+++ packages/frontend/src/redux/slices/sign/index.js
@@ -7,8 +7,6 @@ import { Mixpanel } from '../../../mixpanel';
 import Wallet, { wallet } from '../../../utils/wallet';
 import { showCustomAlert } from '../../actions/status';
 import { selectAccountId, selectAccountUrlPrivateShard } from '../account';
-import { store } from '../../..';
-import { keysToSnakeCase } from '../../../utils/object';
 
 const SLICE_NAME = 'sign';
 
@@ -93,96 +91,6 @@ export const handleSignTransactions = createAsyncThunk(
                 return false;
             }
         },
-    }
-);
-
-export const transactionsProgress = createAction('transactionsProgress');
-
-export const handleTransactionsExecutor = createAsyncThunk(
-    `${SLICE_NAME}/handleTransactionsExecutor`,
-    async ({ txs }, thunkAPI) => {
-        let res = [];
-        const { dispatch, getState } = thunkAPI;
-        const accountId = selectAccountId(getState());
-        const walletAccount = await wallet.getAccount(accountId);
-
-        const result = {
-            success: true,
-            txHash: '',
-            failReason: '',
-        };
-
-        dispatch(
-            transactionsProgress({
-                txs: txs.map((tx) => ({
-                    ...tx,
-                    transaction: {
-                        receiver_id: tx.receiverId,
-                        signer_id: accountId,
-                        actions: tx.actions.map((action) => {
-                            return Object.entries(action).reduce((acc, [key, value]) => {
-                                return {
-                                    ...acc,
-                                    [key.charAt(0).toUpperCase() + key.slice(1)]:
-                                        keysToSnakeCase(value),
-                                };
-                            }, {});
-                        }),
-                    },
-                })),
-            })
-        );
-        for (let i = 0; i < txs.length; i++) {
-            const tx = txs[i];
-            dispatch(
-                transactionsProgress({
-                    txIndex: i,
-                    txProgress: 'signing',
-                })
-            );
-            const txRes = await walletAccount.signAndSendTransaction(tx);
-            res.push(txRes);
-
-            const {
-                status,
-                receipts_outcome,
-                transaction: { hash },
-            } = txRes;
-            result.txHash = hash;
-
-            const failedResult = receipts_outcome.find(
-                ({ outcome: { status } }) => !!status?.Failure
-            );
-
-            if (failedResult?.outcome?.status?.Failure) {
-                result.success = false;
-                result.failReason = failedResult.outcome.status.Failure;
-                dispatch(
-                    transactionsProgress({
-                        txIndex: i,
-                        hash,
-                        status,
-                        txProgress: 'failed',
-                    })
-                );
-            } else {
-                dispatch(
-                    transactionsProgress({
-                        txIndex: i,
-                        hash,
-                        status,
-                        txProgress: 'success',
-                    })
-                );
-            }
-        }
-        result.data = res;
-        dispatch(
-            transactionsProgress({
-                txs: [],
-            })
-        );
-        return result;
     }
 );
 
@@ -369,21 +277,3 @@ export const selectSignFeesGasLimitIncludingGasChanges = createSelector(
         return new BN(calculateGasLimit(tx.flatMap((t) => t.actions))).add(gasUsed);
     }
 );
-
-export async function dispatchTransactionExecutor(txOption) {
-    const res = await store.dispatch(
-        await handleTransactionsExecutor({
-            txs: [txOption],
-        })
-    );
-    return res.payload.data[0];
-}
-
-export async function dispatchTransactionsExecutor(txs) {
-    const res = await store.dispatch(
-        await handleTransactionsExecutor({
-            txs,
-        })
-    );
-    return res.payload;
-}
