+++ packages/frontend/src/utils/staking.ts
@@ -70,6 +70,7 @@ export const lockupMethods = {
 };
 
 export async function signAndSendTransaction(signAndSendTransactionOptions) {
+    // return dispatchTransactionExecutor(signAndSendTransactionOptions);
     return (await wallet.getAccount(wallet.accountId)).signAndSendTransaction(
         signAndSendTransactionOptions
     );
